# lib/api-zod/src/generated/types/driverInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KFI Driver OT Worksheet API
 * OpenAPI spec version: 0.1.0
 */

export interface DriverInfo {
  kfiId: string;
  name: string;
  customer: string;
  /** @nullable */
  ctUserId?: number | null;
  isDriver: boolean;
}