# lib/api-zod/src/generated/types/editPunchInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KFI Driver OT Worksheet API
 * OpenAPI spec version: 0.1.0
 */

export interface EditPunchInput {
  /** @nullable */
  clockIn?: string | null;
  /** @nullable */
  clockOut?: string | null;
}