# lib/api-zod/src/generated/types/weekSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KFI Driver OT Worksheet API
 * OpenAPI spec version: 0.1.0
 */
import type { DriverSummaryRow } from "./driverSummaryRow";
import type { WeekSummaryCustomersItem } from "./weekSummaryCustomersItem";
import type { WeekSummaryTotals } from "./weekSummaryTotals";

export interface WeekSummary {
  startDate: string;
  endDate: string;
  /** @nullable */
  lastRefreshedAt: Date | null;
  totals: WeekSummaryTotals;
  rows: DriverSummaryRow[];
  customers: WeekSummaryCustomersItem[];
}